OneDrive/Desktop/PROJECT/GitHub Projects/safetysnap-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
Pillow
ultralytics
opencv-python-headless
gunicorn
psycopg2-binary
python-dotenv